make autocompleter abstract
